SPLITT: A generic C++ library for Serial and Parallel Lineage Traversal of Trees

All basic types, functions and classes defined in SPLITT.

Basic types

uint

uvec

vec

bvec

Global constants

G_NA_UINT

G_EMPTY_UVEC

G_PI

Global functions

SortIndices

At

Match

Seq

IsNA

NotIsNA

Classes

TraversalSpecification

TraversalTask

TraversalTaskLightweight

Tree

OrderedTree

ThreadExceptionHandler

[[Rcpp::export]]